feat: created gbfs populate class
