+++ .github/workflows/db-update-dev.yml
@@ -6,7 +6,8 @@ on:
       - main
     paths:
       - 'liquibase/changelog.xml'
-      - 'api/src/scripts/populate_db.py'
+      - 'api/src/scripts/populate_db_gtfs.py'
+      - 'api/src/scripts/populate_db_gbfs.py'
   repository_dispatch: # Update on mobility-database-catalog repo dispatch
     types: [ catalog-sources-updated ]
   workflow_dispatch:
