+++ .github/workflows/db-update-qa.yml
@@ -6,7 +6,8 @@ on:
       - main
     paths:
       - 'liquibase/changelog.xml'
-      - 'api/src/scripts/populate_db.py'
+      - 'api/src/scripts/populate_db_gtfs.py'
+      - 'api/src/scripts/populate_db_gbfs.py'
   workflow_dispatch:
 jobs:
   update:
