+++ api/tests/test_utils/database.py
@@ -8,7 +8,7 @@ from sqlalchemy.engine.url import make_url
 from tests.test_utils.db_utils import dump_database, is_test_db, dump_raw_database, empty_database
 from database.database import Database
 
-from scripts.populate_db import DatabasePopulateHelper
+from scripts.populate_db_gtfs import GTFSDatabasePopulateHelper
 from scripts.populate_db_test_data import DatabasePopulateTestDataHelper
 
 import os
@@ -48,7 +48,7 @@ def populate_database(db: Database, data_dirs: str):
         if len(csv_filepaths) == 0:
             raise Exception("No sources_test.csv file found in test_data directories")
 
-        db_helper = DatabasePopulateHelper(csv_filepaths)
+        db_helper = GTFSDatabasePopulateHelper(csv_filepaths)
         db_helper.initialize(trigger_downstream_tasks=False)
 
         # Make a list of all the extra_test_data.json files in the test_data directories and load the data
